pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "wine_quality_epml"
version = "0.1.0"
description = "ML engineering homework project on wine quality."
requires-python = ">=3.10"
dependencies = [
    "dvc[gs]>=3.64.2",
    "scikit-learn>=1.4.0",
    "luigi>=3.5.0",
    "ruamel.yaml>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.4.10",
    "mypy==1.10.0",
    "pre-commit==3.6.2",
]

[tool.uv]
dev-dependencies = [
    "ruff==0.4.10",
    "mypy==1.10.0",
    "pre-commit==3.6.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "S"]
ignore = ["E203"]
fixable = ["ALL"]

[tool.ruff.format]
